add latlon2utm

+++ include/autocalib/lidar2imu/common/Lidar_parser_base.h
@@ -31,7 +31,7 @@ struct LaserCorrection {
 struct LidarPointXYZIRT {
   PCL_ADD_POINT4D;
   float intensity;
-  uint16_t ring;
+  float ring;
   double timestamp;
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 } EIGEN_ALIGN16;
@@ -39,4 +39,4 @@ struct LidarPointXYZIRT {
 POINT_CLOUD_REGISTER_POINT_STRUCT(
     LidarPointXYZIRT,
     (float, x, x)(float, y, y)(float, z, z)(float, intensity, intensity)(
-        uint16_t, ring, ring)(double, timestamp, timestamp))
+      float, ring, ring)(double, timestamp, timestamp))
